django/requirements.txt
django==4.2.5
flake8==6.1.0
python-dotenv==1.0.0
django-split-settings==1.2.0
psycopg2-binary==2.9.9
uwsgi==2.0.23
djangorestframework==3.14.0
django-debug-toolbar==4.2.0
django-cors-headers==4.3.0